// src/app/Models/societe/SocieteG.ts
export class SocieteG {
  societeId: number;
  companyId: number;
  raisonsocial: string;
  formjuridique: string;
  abreviation: string;
  adresse: string;
  ville: string;
  pays: string;
  tele: string;
  fax: string;
  tvadebit: number;
  tvaEncai: number;
  tvaigrs: number;
  tvaigrn: number;
  tvaisn: number;
  tvaiss: number;
  rc: string;
  patente: string;
  cnss: string;
  iff: string;
  ice: string;
}